refactor(widgets): CalendarWidget sous 300 lignes — éclatement complet (H8)
- composable useCalendarWidget (état + config FullCalendar + handlers + navigation)
- sous-composants présentationnels CalendarWidgetLegend + CalendarEventModal (CSS verbatim)
- orchestrateur 570→258 lignes
- tests: composable (5) + modale (3) + légende (1)
- parité: mêmes options FullCalendar, mêmes classes/couleurs, formatEventTime inchangé
